+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # CHANGELOG
 
+## 0.0.5 - 29.05.2025
+- OJP 1.0: adds `TripInformationRequest` response model - [PR #8](https://github.com/openTdataCH/ojp-shared-types/pull/8)
+- OJP 1.0: adds `Extension` for TripRequest TimedLeg response
+- harmonize types names
+
+## 0.0.4 - 25.05.2025
+- **BREAKING CHANGE:** dont use `OpenAPI_Dependencies.MapParentArrayTags` anymore, that will be dynamically created in the parser
+- adds `TripInformationRequest` request / response models
+- make `ResponseContext`, `LegTrack / TrackSectionStructure` shared
+
 ## 0.0.2 - 7.05.2025
 - created repo, initial commit